+++ .github/checks.yml
@@ -27,4 +27,4 @@ jobs:
       - name: Install dependencies
         run: pnpm install
       - name: Run checks
-        run: pnpm run::checks
+        run: pnpm run:checks
