+++ package.json
@@ -12,9 +12,9 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint .",
-		"gen::readme": "ruby ./scripts/gen_readme.rb",
-		"gen::palette": "ruby ./scripts/gen_colors.rb",
-		"run::checks": "ruby ./scripts/checks.rb"
+		"gen:readme": "ruby ./scripts/gen_readme.rb",
+		"gen:palette": "ruby ./scripts/gen_colors.rb",
+		"run:checks": "ruby ./scripts/checks.rb"
 	},
 	"devDependencies": {
 		"@eslint/compat": "^2.0.0",
@@ -29,7 +29,6 @@
 		"globals": "^16.5.0",
 		"prettier": "^3.6.2",
 		"prettier-plugin-svelte": "^3.4.0",
-		"sass": "^1.94.1",
 		"svelte": "^5.43.12",
 		"svelte-check": "^4.3.4",
 		"typescript": "^5.9.3",
