+++ palette/palette.scss
@@ -1,25 +1,25 @@
 :root {
-  --base: #181822;
-  --pink: #d35c9e;
-  --text: #c9d3ea;
-  --subtext0: #a3aec9;
-  --subtext1: #7d89a6;
-  --purple: #8c7cf0;
-  --crust: #0c0b11;
-  --mantle: #12121a;
-  --sapphire: #79d6fc;
-  --surface0: #2a2937;
-  --surface1: #353546;
-  --surface2: #403f55;
-  --overlay0: #4a4963;
-  --overlay1: #55546f;
-  --overlay2: #605f7a;
-  --yellow: #e6d163;
-  --peach: #e89a7a;
-  --green: #68c792;
-  --blue: #2a7dd3;
-  --orange: #e0853e;
-  --sky: #4ebbe6;
-  --red: #e04f5f;
-  --black: #05070b;
+	--base: #181822;
+	--pink: #d35c9e;
+	--text: #c9d3ea;
+	--subtext0: #a3aec9;
+	--subtext1: #7d89a6;
+	--purple: #8c7cf0;
+	--crust: #0c0b11;
+	--mantle: #12121a;
+	--sapphire: #79d6fc;
+	--surface0: #2a2937;
+	--surface1: #353546;
+	--surface2: #403f55;
+	--overlay0: #4a4963;
+	--overlay1: #55546f;
+	--overlay2: #605f7a;
+	--yellow: #e6d163;
+	--peach: #e89a7a;
+	--green: #68c792;
+	--blue: #2a7dd3;
+	--orange: #e0853e;
+	--sky: #4ebbe6;
+	--red: #e04f5f;
+	--black: #05070b;
 }
