+++ src/lib/components/global/Footer.svelte
