+++ src/lib/components/home/About.svelte
