+++ src/lib/components/home/Media.svelte
