+++ src/lib/components/home/Projects.svelte
