+++ src/lib/components/home/Tech.svelte
