+++ src/lib/components/moons/FirstQuarter.svelte
