+++ src/lib/components/moons/NewMoon.svelte
