+++ src/lib/components/moons/WaningCrescent.svelte
