+++ src/lib/components/moons/WaningGibbous.svelte
