+++ src/lib/components/moons/WaxingCrescent.svelte
