+++ src/lib/components/moons/WaxingGibbous.svelte
