+++ src/routes/+layout.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import '../styles/main.scss';
 	let { children } = $props();
 </script>
