+++ src/routes/+page.svelte
@@ -1,24 +1,4 @@
 <script lang="ts">
-	import Tech from '../components/home/Tech.svelte';
-	import Top from '../components/home/Top.svelte';
-	import Footer from '../components/global/Footer.svelte';
-	import About from '../components/home/About.svelte';
-	import Media from '../components/home/Media.svelte';
-	import Projects from '../components/home/Projects.svelte';
-	import GitHub from '../components/home/GitHub.svelte';
 </script>
 
-<div class="home-wrapper">
-	<Top />
-	<hr />
-	<About />
-	<hr />
-	<Tech />
-	<hr />
-	<Projects />
-	<hr />
-	<GitHub />
-	<hr />
-	<Media />
-	<Footer />
-</div>
+<div class="home-wrapper"></div>
