--- src/styles/global/footer.scss
@@ -1,12 +1,0 @@
-@use 'mixins';
-
-//footer styling
-.footer-container {
-	@include mixins.flex(center, center, column);
-	select {
-		margin-top: 5px;
-		background-color: var(--mantle);
-		color: var(--text);
-		border-color: var(--crust);
-	}
-}
