--- src/styles/global/mixins.scss
@@ -1,24 +1,0 @@
-@mixin flex($align: center, $justify: center, $direction: row) {
-	display: flex;
-	align-items: $align;
-	justify-content: $justify;
-	flex-direction: $direction;
-}
-
-@mixin md($width, $height) {
-	& {
-		@media (width >= 1024px) {
-			width: $width;
-			height: $height;
-		}
-	}
-}
-
-@mixin sm($width, $height) {
-	& {
-		@media (width < 1024px) {
-			width: $width;
-			height: $height;
-		}
-	}
-}
