--- src/styles/home/home.scss
@@ -1,102 +1,0 @@
-@use '../global/mixins';
-@use '../global/globals';
-@use 'sections/media';
-@use 'sections/palette';
-@use 'sections/projects';
-@use 'sections/tech';
-@use 'sections/favourites.scss';
-
-footer {
-  text-align: center;
-}
-
-h2 {
-  text-align: left;
-
-  @media only screen and (max-width: 700px) {
-    text-align: center;
-  }
-}
-
-.home-wrapper {
-  min-width: 300px;
-  width: 55%;
-  margin: 2% auto 0 auto;
-}
-
-.top-content {
-  @include mixins.flex(flex-start, center, column);
-  padding-left: 1.5rem;
-
-  .avatar-name {
-    @include mixins.flex(center, flex-start);
-
-    img {
-      margin-right: 10px;
-      border-radius: 50%;
-    }
-  }
-
-  .spotify-div {
-    display: flex;
-    align-items: center;
-    color: var(--text);
-    font-size: clamp(12px, 2vw, 16px);
-    text-decoration-line: none;
-    margin-top: 10px;
-    text-align: start;
-
-    img {
-      margin-left: 7px;
-      margin-right: 5px;
-    }
-
-    @media only screen and (max-width: 700px) {
-      flex-direction: column;
-      text-align: center;
-      align-items: center;
-
-      img {
-        width: 60px;
-        height: 60px;
-        margin-top: 10px;
-      }
-    }
-
-    .spotify-content {
-      margin-left: 2px;
-
-      @media only screen and (max-width: 700px) {
-        margin-top: 10px;
-      }
-    }
-  }
-
-  @media only screen and (max-width: 700px) {
-    padding-left: 0;
-    align-items: center;
-  }
-
-  h2 {
-    margin-bottom: 0;
-  }
-}
-
-hr {
-  border-color: var(--sapphire);
-  margin: 1.5rem 0 1.5rem 0;
-}
-
-.about-content,
-.github-content {
-  text-align: left;
-  padding: 0 1.5rem;
-
-  h2 {
-    margin-bottom: 10px;
-  }
-
-  @media only screen and (max-width: 700px) {
-    text-align: center;
-  }
-}
