--- src/styles/home/sections/media.scss
@@ -1,17 +1,0 @@
-@use '../../global/mixins';
-
-.media-content {
-  @include mixins.flex();
-  margin-bottom: 10px;
-  margin-top: 1.5rem;
-
-  a {
-    width: 50px;
-    fill: var(--sapphire);
-    transition: fill 0.5s;
-
-    &:hover {
-      fill: var(--overlay2);
-    }
-  }
-}
