--- src/styles/home/sections/projects.scss
@@ -1,69 +1,0 @@
-@use '../../global/mixins';
-@use '../../global/globals';
-
-.projects-wrapper {
-  h2 {
-    margin-bottom: 1rem;
-    padding-left: 1.5rem;
-
-    @media only screen and (max-width: 700px) {
-      padding-left: 0;
-    }
-  }
-}
-
-.projects-content {
-  margin: 0 1.5rem;
-  display: grid;
-  grid-template-columns: repeat(2, minmax(300px, 1fr));
-  gap: 1rem;
-
-  @media only screen and (max-width: 700px) {
-    margin: 0.5rem;
-  }
-
-  @media only screen and (max-width: 1240px) {
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  }
-}
-
-.project {
-  background-color: var(--crust);
-  border: 2px solid var(--sapphire);
-  border-radius: 12px;
-  aspect-ratio: auto;
-  padding: 0.75rem 15px;
-  margin: 0;
-  min-height: 120px;
-  @include mixins.flex(center, space-between, column);
-
-  img {
-    max-width: 100%;
-    height: auto;
-  }
-
-  a {
-    color: var(--sapphire);
-    font-size: clamp(18px, 3vw, 20px);
-    text-decoration: none;
-
-    &:hover {
-      color: var(--overlay2);
-    }
-  }
-
-  article {
-    font-size: clamp(14px, 2vw, 16px);
-
-    a {
-      font-size: clamp(14px, 2vw, 16px);
-    }
-  }
-
-  .project-top-content {
-    @include mixins.flex(flex-start, space-evenly, column);
-    height: 70%;
-    width: 100%;
-    margin-left: 2%;
-  }
-}
