--- src/styles/main.scss
@@ -1,3 +1,0 @@
-@use 'global/globals';
-@use 'home/home.scss';
-@use '../../palette/palette';
